djangorest/requirements.txt
django==4.0.7
djangorestframework==3.13.1
django-stubs==1.12.0
django-configurations[cache,database,email,search]==2.3.2
django-filter==22.1
drf-yasg==1.21.3
djangorestframework-simplejwt==5.2.0
PyJWT==2.4.0
Pillow==9.2.0
celery==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.3.0
redis==4.3.4
eventlet==0.33.1
pyyaml==6.0
uritemplate==4.1.1
psycopg2==2.9.3
dj-database-url==1.0.0
coverage==6.4.4
beautifulsoup4==4.11.1
docutils==0.19
uWSGI==2.0.20